Improve workspace/didChangeConfiguration behavior
I realized that the new settings are passed, so there is no need to send configuration request there.
