+++ lib/rucoa/handler_concerns.rb
@@ -2,7 +2,6 @@
 
 module Rucoa
   module HandlerConcerns
-    autoload :ConfigurationRequestable, 'rucoa/handler_concerns/configuration_requestable'
     autoload :DiagnosticsPublishable, 'rucoa/handler_concerns/diagnostics_publishable'
   end
 end
