+++ lib/rucoa/handlers/initialized_handler.rb
@@ -3,7 +3,7 @@
 module Rucoa
   module Handlers
     class InitializedHandler < Base
-      include HandlerConcerns::ConfigurationRequestable
+      include HandlerConcerns::DiagnosticsPublishable
 
       def call
         request_workspace_configuration
@@ -17,6 +17,23 @@ module Rucoa
         YardGlobDocumentLoader.call(
           globs: ::YARD::Parser::SourceParser::DEFAULT_PATH_GLOB
         )
+      end
+
+      # @return [void]
+      def request_workspace_configuration
+        write(
+          method: 'workspace/configuration',
+          params: {
+            items: [
+              {
+                section: 'rucoa'
+              }
+            ]
+          }
+        ) do |response|
+          configuration.update(response['result'][0])
+          publish_diagnostics_on_each_source
+        end
       end
     end
   end
